requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
matplotlib>=3.5.0
scikit-image>=0.19.0
Pillow>=9.0.0
tqdm>=4.64.0
tensorboard>=2.9.0
pyyaml>=6.0
opencv-python>=4.6.0
scipy>=1.8.0
seaborn>=0.11.0
wandb>=0.13.0
jupyterlab>=3.4.0
ipywidgets>=7.7.0